move buttonStyles into styles
