+++ src/core/components/button/Button.tsx
@@ -1,12 +1,7 @@
-import { ReactElement, ReactNode, ButtonHTMLAttributes } from 'react';
-import { SerializedStyles } from '@emotion/react';
-import { buttonStyles, buttonContents } from './shared';
-import type {
-	SharedButtonProps,
-	ButtonPriority,
-	Size,
-	IconSide,
-} from './types';
+import { ButtonHTMLAttributes } from 'react';
+import { buttonContents } from './shared';
+import { buttonStyles } from './styles';
+import type { SharedButtonProps } from './types';
 
 export interface ButtonProps
 	extends SharedButtonProps,
